Rollback mime-types change because path package is polyfilled and convert url.js to Typescript
